requirements.txt
gunicorn==19.6.0
Flask==1.0.2
Flask-Compress==1.4.0
gevent==1.4.0
gym==0.10.5
h5py==2.9.0
html5lib==1.0.1
idna==2.7
image==1.5.27
imageio==2.5.0
Jinja2==2.10
jmespath==0.9.3
joblib==0.13.0
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
numpy==1.16.1
Pillow==6.0.0
requests==2.19.1
requests-oauthlib==1.2.0
scikit-image==0.15.0
scikit-learn==0.19.2
scipy==1.1.0
tensorflow==2.0.0a0
tensorflow-estimator==1.14.0
testpath==0.3.1
tf-estimator-nightly==1.14.0.dev2019030115
Werkzeug==0.14.1

